+++ src/ufmt.rs
@@ -1,8 +1,19 @@
 use crate::{string::String, vec::Vec};
 use ufmt_write::uWrite;
 
+impl<const SIZE: usize> uDisplay for String<SIZE> {
+    #[inline(always)]
+    fn fmt<W>(&self, f: &mut ufmt::Formatter<'_, W>) -> Result<(), W::Error>
+    where
+        W: uWrite + ?Sized,
+    {
+        f.write_str(&self.as_str())
+    }
+}
+
 impl<const N: usize> uWrite for String<N> {
     type Error = ();
+    #[inline(always)]
     fn write_str(&mut self, s: &str) -> Result<(), Self::Error> {
         self.push_str(s)
     }
@@ -10,6 +21,7 @@ impl<const N: usize> uWrite for String<N> {
 
 impl<const N: usize> uWrite for Vec<u8, N> {
     type Error = ();
+    #[inline(always)]
     fn write_str(&mut self, s: &str) -> Result<(), Self::Error> {
         self.extend_from_slice(s.as_bytes())
     }
@@ -28,7 +40,16 @@ mod tests {
     }
 
     #[test]
-    fn test_string() {
+    fn test_udisplay_string() {
+        let str_a = String::<32>::from("world");
+        let str_b = String::<32>::new();
+        uwrite!(str_b, "Hello {}!", str_a).unwrap();
+
+        assert_eq!(str_b, "Hello world!");
+    }
+
+    #[test]
+    fn test_uwrite_string() {
         let a = 123;
         let b = Pair { x: 0, y: 1234 };
 
@@ -39,14 +60,14 @@ mod tests {
     }
 
     #[test]
-    fn test_string_err() {
+    fn test_uwrite_string_err() {
         let p = Pair { x: 0, y: 1234 };
         let mut s = String::<4>::new();
         assert!(uwrite!(s, "{:?}", p).is_err());
     }
 
     #[test]
-    fn test_vec() {
+    fn test_uwrite_vec() {
         let a = 123;
         let b = Pair { x: 0, y: 1234 };